requirements.txt
Flask
Flask-Bootstrap4
Flask-Moment
Flask-WTF
jinja2
WTForms
requests
Werkzeug
hdbcli
